+++ backend/canisters/community/api/src/queries/api_key.rs
@@ -17,4 +17,5 @@ pub enum Response {
     NotAuthorized,
     NotFound,
     ChannelNotFound,
+    CommunityFrozen,
 }