+++ backend/canisters/group/api/src/queries/api_key.rs
@@ -13,6 +13,7 @@ pub struct Args {
 #[derive(Serialize, Deserialize, Debug)]
 pub enum Response {
     Success(ApiKey),
+    GroupFrozen,
     NotAuthorized,
     NotFound,
 }